20240426- v3.24 Fixed Multiline texts notes
20240426- v3.24 Fixed Multiline texts notes
Estimated Time: 0
Estimated Total Time: 32 : 25

=Bugs Squashed====================================================================================================
--20240426-0m |Now notes are not getting new lines....
	Notes: Fixed the Kanban tab loading page first bug and now the notes dont add new lines correctly
I dont
		know why...
New notes typed directly in DO have the new lines. Loaded notes do not.
But in the
		TodoItemEditor, they ARE getting new lines....
Adding a new line in the editor makes the notes panel update with
		new lines...maybe it's not getting refreshed?
Problem: I don't know. It used to work, and then it
		doesn't. Neither does it on older versions that it did work. Maybe something changed in
		Windows?!?
Solution: On setting the TextBox.Text property, I replace the "\n" with Environment.NewLine and it works now